fixes #3312
